permessi

+++ templates/home.html.php
@@ -12,14 +12,16 @@
             <div class="body">
                 <p><?=$post['body'] ?></p>
             </div>
-            <div class="post-edit">
-                <a class="edit-link" href="index.php?route=post/edit&id=<?=$post['id']?>">
-                    edit
-                </a>
-                <a class="delete-link" href="index.php?route=post/delete&id=<?=$post['id']?>">
-                    <i class="fas fa-trash-alt"></i>
-                </a>
-            </div>
+            <?php if ($post['user_id'] == $user['id']) : ?>
+                <div class="post-edit">
+                    <a class="edit-link" href="index.php?route=post/edit&id=<?=$post['id']?>">
+                        edit
+                    </a>
+                    <a class="delete-link" href="index.php?route=post/delete&id=<?=$post['id']?>">
+                        <i class="fas fa-trash-alt"></i>
+                    </a>
+                </div>
+            <?php endif; ?>
         </div>
     <?php endforeach; ?>
 </section>
